requirements.txt
ipython
protobuf
python-dotenv
streamlit==1.51.0
watchdog==6.0.0
google-adk==1.19.0
sqlalchemy[asyncio]
nest_asyncio